cpp-rule-preprocessor receives one file at a time
