+++ cpp2rust/cpp2rust.cpp
@@ -3,14 +3,24 @@
 
 #include <llvm/Support/CommandLine.h>
 
+#include <algorithm>
+#include <array>
 #include <cstdlib>
 #include <filesystem>
 #include <fstream>
 #include <vector>
 
+#if defined(_WIN32)
+#include <windows.h>
+#elif defined(__linux__)
+#include <limits.h>
+#include <unistd.h>
+#elif defined(__APPLE__)
+#include <mach-o/dyld.h>
+#endif
+
 #include "cpp2rust_lib.h"
 #include "logging.h"
-#include "rules_dir.h"
 
 namespace fs = std::filesystem;
 
@@ -55,6 +65,41 @@ llvm::cl::list<std::string> CXXFlags("cxxflags",
                                      llvm::cl::cat(cpp2rust_cmdargs));
 
 } // namespace
+
+static fs::path GetExecutableDir() {
+#if defined(_WIN32)
+  char path[MAX_PATH];
+  GetModuleFileNameA(NULL, path, MAX_PATH);
+  return fs::path(path).parent_path();
+#elif defined(__linux__)
+  char path[PATH_MAX];
+  ssize_t count = readlink("/proc/self/exe", path, PATH_MAX);
+  return fs::path(std::string_view(path, std::max((ssize_t)0, count)))
+      .parent_path();
+#elif defined(__APPLE__)
+  uint32_t size = 0;
+  _NSGetExecutablePath(nullptr, &size);
+  std::vector<char> buffer(size);
+  _NSGetExecutablePath(buffer.data(), &size);
+  return fs::path(buffer.data()).parent_path();
+#endif
+  return ".";
+}
+
+static bool ResolveRulesDir() {
+  std::array<fs::path, 3> candidates = {fs::path("./rules"),
+                                        fs::path("../rules"),
+                                        GetExecutableDir() / "../../rules"};
+
+  for (const auto &dir : candidates) {
+    if (fs::exists(dir) && fs::is_directory(dir)) {
+      RulesDir = fs::canonical(dir).string();
+      llvm::errs() << "Using rules directory: " << RulesDir << '\n';
+      return true;
+    }
+  }
+  return false;
+}
 
 int main(int argc, char *argv[]) {
   llvm::cl::HideUnrelatedOptions(cpp2rust_cmdargs);
@@ -104,7 +149,9 @@ int main(int argc, char *argv[]) {
 
   std::vector<std::string_view> cxx_flags(CXXFlags.begin(), CXXFlags.end());
 
-  if (RulesDir.empty() && !cpp2rust::ResolveRulesDir(RulesDir)) {
+  if (RulesDir.empty() && !ResolveRulesDir()) {
+    llvm::errs() << "ERROR: rules directory not found. "
+                    "Please specify one with --rules\n";
     return EXIT_FAILURE;
   }
